requirements.txt
langgraph>=0.0.40,
langchain>=0.1.20,
langchain-google-vertexai>=1.0.1,
langchain-community>=0.0.25,
python-dotenv>=1.0.1,